TapAutomationProject/requirements.txt
streamlit
SpeechRecognition
requests
python-dotenv
google-generativeai 